requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
boto3==1.34.32
botocore==1.34.32
bs4==0.0.2
cachetools==5.3.3
certifi==2023.11.17
click==8.1.7
dashscope==1.16.0
dgl==2.0.0
et-xmlfile==1.1.0
fire==0.6.0
flanker==0.9.11
fonttools==4.25.0
frozenlist==1.4.1
fsspec==2023.12.2
funcy==2.0
gensim==4.3.2
google==3.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
htmlfilter==0.2
huggingface-hub==0.21.3
jieba==0.42.1
joblib==1.3.2
jsonpatch==1.33
jsonpath==0.82.2
jsonpointer==2.4
matplotlib==3.5.1
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.5
munkres==1.1.4
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openai==0.28.0
openpyxl==3.1.0
outcome==1.3.0.post0
pandas==2.2.0
ply==3.11
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
pyarrow==15.0.0
pyasn1-modules==0.2.8
PyDispatcher==2.0.5
pyLDAvis==3.4.1
PyQt5==5.15.10
PySnooper==1.2.0
pytorch-crf==0.7.2
pytorch-pretrained-bert==0.6.2
pywin32==305.1
PyYAML==6.0.1
regex==2023.12.25
rsa==4.9
rules==3.3
s3transfer==0.10.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
selenium==4.19.0
smart-open==6.4.0
sortedcontainers==2.4.0
sympy==1.12
tensorboardX==2.6.2.2
termcolor==2.4.0
textblob==0.18.0.post0
threadpoolctl==3.2.0
tld==0.13
tokenizers==0.15.2
torchdata==0.7.1
torchnet==0.0.4
TorchSnooper==0.8
transformers==4.38.1
trio==0.25.0
trio-websocket==0.11.1
urllib3==1.26.18
visdom==0.2.4
WebOb==1.8.7
websocket-client==1.7.0
wsproto==1.2.0
xarray==0.20.1
yarl==1.9.4
